# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FOODORO POS & Kitchen Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "404":
          description: Not found
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ─── Products ─────────────────────────────────────────────────────────────
...
  /kitchen/tickets:
    get:
      operationId: listKitchenTickets
      tags: [kitchen]
      summary: List kitchen tickets (active orders for KDS)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [new, in_progress, ready]
      responses:
        "200":
          description: Kitchen tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KitchenTicket"
  /kitchen/tickets/{id}/status:
    patch:
      operationId: updateTicketStatus
      tags: [kitchen]
      summary: Update kitchen ticket status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KitchenTicket"
        "404":
          description: Not found
...
  /kitchen/availability/{productId}:
    patch:
      operationId: setProductAvailability
      tags: [kitchen]
      summary: Set kitchen availability for a product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetProductAvailabilityBody"
      responses:
        "200":
          description: Updated availability
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductAvailabilityItem"
        "404":
          description: Product not found

  # ─── Inventory ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Category:
      type: object
      required: [id, name, color, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        color:
          type: string
        createdAt:
          type: string
          format: date-time
    CategoryInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        color:
          type: string
    CategoryUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        color:
          type: string

    # Products
...
    OrderItem:
      type: object
      required: [id, orderId, productId, productName, quantity, unitPrice, subtotal]
      properties:
        id:
          type: integer
        orderId:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        subtotal:
          type: number
        notes:
          type: ["string", "null"]
...
    KitchenTicket:
      type: object
      required: [id, orderId, orderNumber, status, type, items, createdAt]
      properties:
        id:
          type: integer
        orderId:
          type: integer
        orderNumber:
          type: string
        status:
          type: string
          enum: [new, in_progress, ready, completed]
        type:
          type: string
          enum: [dine_in, takeaway, delivery]
        tableNumber:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TicketStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [new, in_progress, ready, completed]

    # Product Ingredients
...
    UnavailabilityReason:
      type: string
      enum: [out_of_stock, temp_unavailable, ended_today, ingredient_out, paused]
    ProductAvailabilityItem:
      type: object
      required: [id, name, price, isActive, categoryId, kitchenAvailable]
      properties:
        id:
          type: integer
        name:
          type: string
        price:
          type: number
        isActive:
          type: boolean
        categoryId:
          type: integer
        kitchenAvailable:
          type: boolean
        unavailabilityReason:
          type: ["string", "null"]
        unavailableUntil:
          type: ["string", "null"]
          format: date-time
          description: If set, the product will be automatically re-enabled at this time
    SetProductAvailabilityBody:
      type: object
      required: [available]
      properties:
        available:
          type: boolean
        reason:
          $ref: "#/components/schemas/UnavailabilityReason"
        reasonNote:
          type: string
        changedBy:
          type: string
        unavailableUntil:
          type: ["string", "null"]
          format: date-time
          description: ISO 8601 timestamp — auto re-enables the product at this time
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT